requirements.txt
openai>=1.30.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pandas>=2.2.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
pytest>=8.2.0
